コントローラ、モデルのテストコードの記述

+++ app/controllers/v1/users_controller.rb
@@ -26,7 +26,8 @@ class V1::UsersController < ApplicationController
   def update
     @user = User.find(params[:id])
 
-    if params[:profile_image] == nil || params[:file_name] == nil
+    # ユーザがプロフィール画像を削除したら cloudinary に保存してある画像も削除する
+    if params[:profile_image] == nil && params[:file_name] == nil
       Cloudinary::Api.delete_resources([@user.file_name]) if @user.file_name
     end
